refactor(webhook): split check_runs.go by responsibility
Move the 1500+ line pkg/webhook/check_runs.go into four focused files
within the same package. Pure file-organization refactor — no logic,
signature, or visibility changes.

- check_runs.go: constants, types, and shared header
- check_aggregate.go: aggregate check naming, filtering, conclusion derivation
- check_records.go: per-database check record storage and reconciliation
- check_prior_env.go: prior-environment ordering enforcement
- check_publisher.go: GitHub Check Run publishing and SHA-current verification

Validation:
- go build ./...
- go test -count=1 -race ./pkg/webhook/...
- make test-unit (all packages pass)
- golangci-lint run --new-from-rev=origin/main ./pkg/webhook/... (0 issues)
